CAMEL-20355: refactor the throttle tests
Adjust code to facilitate accommodating more tests for throttling
