+++ components/camel-spring-xml/src/test/java/org/apache/camel/spring/processor/throttle/SpringThrottlerGroupingTest.java
@@ -14,14 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.spring.processor;
+package org.apache.camel.spring.processor.throttle;
 
 import java.util.concurrent.Semaphore;
 
 import org.apache.camel.CamelContext;
 import org.apache.camel.Exchange;
 import org.apache.camel.Processor;
-import org.apache.camel.processor.ThrottlingGroupingTest;
+import org.apache.camel.processor.throttle.ThrottlingGroupingTest;
 
 import static org.apache.camel.spring.processor.SpringTestHelper.createSpringCamelContext;
 import static org.junit.jupiter.api.Assertions.assertTrue;
