+++ components/camel-spring-xml/src/test/java/org/apache/camel/spring/processor/throttle/SpringThrottlerThreadPoolProfileTest.java
@@ -14,24 +14,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.processor;
+package org.apache.camel.spring.processor.throttle;
 
-public class ThrottlingException extends RuntimeException {
+import org.apache.camel.CamelContext;
+import org.apache.camel.processor.throttle.ThrottlerThreadPoolProfileTest;
 
-    private static final long serialVersionUID = 1993185881371058773L;
+import static org.apache.camel.spring.processor.SpringTestHelper.createSpringCamelContext;
 
-    public ThrottlingException() {
-    }
+public class SpringThrottlerThreadPoolProfileTest extends ThrottlerThreadPoolProfileTest {
 
-    public ThrottlingException(String message) {
-        super(message);
-    }
-
-    public ThrottlingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ThrottlingException(Throwable cause) {
-        super(cause);
+    @Override
+    protected CamelContext createCamelContext() throws Exception {
+        return createSpringCamelContext(this,
+                "org/apache/camel/spring/processor/ThrottlerThreadPoolProfileTest.xml");
     }
 }
