+++ components/camel-spring-xml/src/test/java/org/apache/camel/spring/processor/throttle/SpringThrottlingInflightRoutePolicyTest.java
@@ -14,10 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.spring.processor;
+package org.apache.camel.spring.processor.throttle;
 
 import org.apache.camel.CamelContext;
-import org.apache.camel.processor.ThrottlingInflightRoutePolicyTest;
+import org.apache.camel.processor.throttle.ThrottlingInflightRoutePolicyTest;
 
 import static org.apache.camel.spring.processor.SpringTestHelper.createSpringCamelContext;
