+++ core/camel-core/src/test/java/org/apache/camel/processor/throttle/ThrottlingExceptionRoutePolicyHalfOpenTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.processor;
+package org.apache.camel.processor.throttle;
 
 import java.util.Arrays;
 import java.util.List;
@@ -26,7 +26,6 @@ import org.apache.camel.Processor;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.apache.camel.support.service.ServiceSupport;
-import org.apache.camel.throttling.ThrottlingExceptionHalfOpenHandler;
 import org.apache.camel.throttling.ThrottlingExceptionRoutePolicy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -37,8 +36,8 @@ import org.slf4j.LoggerFactory;
 import static org.awaitility.Awaitility.await;
 
 @Isolated("Does not play well with parallel execution")
-public class ThrottlingExceptionRoutePolicyHalfOpenHandlerTest extends ContextTestSupport {
-    private static Logger log = LoggerFactory.getLogger(ThrottlingExceptionRoutePolicyHalfOpenHandlerTest.class);
+public class ThrottlingExceptionRoutePolicyHalfOpenTest extends ContextTestSupport {
+    private static Logger log = LoggerFactory.getLogger(ThrottlingExceptionRoutePolicyHalfOpenTest.class);
 
     private String url = "direct:start?block=false";
     private MockEndpoint result;
@@ -55,6 +54,7 @@ public class ThrottlingExceptionRoutePolicyHalfOpenHandlerTest extends ContextTe
 
     @Test
     public void testHalfOpenCircuit() throws Exception {
+        result.reset();
         result.expectedMessageCount(2);
         List<String> bodies = Arrays.asList("Message One", "Message Two");
         result.expectedBodiesReceivedInAnyOrder(bodies);
@@ -110,20 +110,10 @@ public class ThrottlingExceptionRoutePolicyHalfOpenHandlerTest extends ContextTe
                 long halfOpenAfter = 250;
                 ThrottlingExceptionRoutePolicy policy
                         = new ThrottlingExceptionRoutePolicy(threshold, failureWindow, halfOpenAfter, null);
-                policy.setHalfOpenHandler(new AlwaysCloseHandler());
 
-                from(url).routeId("foo").routePolicy(policy).log("${body}").to("log:foo?groupSize=10").to("mock:result");
+                from(url).routeId("foo").routePolicy(policy).to("log:foo?groupSize=10").to("mock:result");
             }
         };
-    }
-
-    public static class AlwaysCloseHandler implements ThrottlingExceptionHalfOpenHandler {
-
-        @Override
-        public boolean isReadyToBeClosed() {
-            return true;
-        }
-
     }
 
     protected void sendMessage(String bodyText) {
