+++ core/camel-core/src/test/java/org/apache/camel/processor/throttle/ThrottlingGroupingTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.processor;
+package org.apache.camel.processor.throttle;
 
 import java.util.HashMap;
 import java.util.Map;
